Bug fix for https://github.com/greenmanspirit/journal.txt/issues/3
I changed the behavior of journal.txt file creation to put no contents in the file. This was to create a zero length file since having a blank line as contents creates an invalid file. I can easily detect a zero length file elsewhere to handle a new file properly.

Then to fix the bug, if the file is not zero length, I check the date of the first entry (should be second line). If
the date is today, I pass work off to the edit_entry method. Otherwise, it rewinds the file and goes on creating a ne
w journal entry.