Enhancement for https://github.com/greenmanspirit/journal.txt/issues/2
Added delete_entry to the journal object. It simply finds the entry and disca
rds it. I also added delete to the usage message, the actions case statement and the README. This is a bit messy but refactoring will come in another commit.  I also changed the edit functionality. Since you can edit to an empty file, the script detects that now and asks if you want to delete the message, if not, then leave the entry alone.