Bug fix for https://github.com/greenmanspirit/journal.txt/issues/16
First I pulled the date checking code into it's own support function so that I could have the fix in one place. It appears that it was as simple as changing the conversion from using Integer to using .to_i as it defaults to base 10. In addition to this, I noticed during testing that delete_entry had a typo in one of it's calls to write_file. I fixed that and edit and delete behave as expected and desired.