Bug/Enhancement for https://github.com/greenmanspirit/journal.txt/issues/11
Even though https://github.com/greenmanspirit/journal.txt/issues/17 was added later, this fix was a lot easier after doing 17. Part of solving this was to make all date variables date objects. This gave me the ability to use '<' to test when we passed the date we want. As a result I had to tweak a few things to get the date objects working right in place of date strings. I have a feeling this will help down the line with adding date formats.