Enhancment for https://github.com/greenmanspirit/journal.txt/issues/4
This was trivially easy to do so I decided to add the ability to filter dates to list as well. I loop through the file and on each date line, I check the date against the filter and if it passes, it is printed to the screen. The filter can be ommited and a default permissive regex is built to test the date, otherwise the filter can use a mix of two digit number and the * wildcard: 01/*/14 for all January 2014 entries. I also updated the usage and readme to reflect this added action.