Bug fix for https://github.com/greenmanspirit/journal.txt/issues/19
In looking at the code, I noticed that when getting the date parts, I started on index 1 instead of 0 so the failure was coming because I was trying to do a valid date for month 18. I then played in irb and it appears that when accessing an array index that was not set, and doing .to_i on it gives you 0. This made a 0 year in the Date.valid_date? call and yeah, year 0 is valid. Gotta love bugs.

Bug fix for https://github.com/greenmanspirit/journal.txt/issues/18

I really need to quit being excited and test more thoroughly. I missed changing @now to Time.now in edit.