Enhancement for https://github.com/greenmanspirit/journal.txt/issues/20
I have created an initial testing suite.
test - This is a file containing all the the tests that need to be run
test_functions - This file contains all the utility functions I created for this
vim - This is a fake vim that writes whatever is in $VALUE to the given file

I will revisit these files later and comment them a bit better. It is late and I want this commited and pushed since it took a bit of thought to work out and I don't trust my laptop.