Followup for https://github.com/greenmanspirit/journal.txt/issues/20
Changed echo to printf so that this vim can create files with no content so I can test saving an empty file. Echo put a newline in, printf does not. I also put in a flag of no_write in VALUE to simulate opening a new file and quitting without saving anything.