Enhancement for https://github.com/greenmanspirit/journal.txt/issues/7
Initial work on this enhancment. I have it detecting and creating if needed the .journalrc file in $HOME. Most of there test.sh files just have an added line at the top creating .journalrc so that there is not query for it. I also needed to edit some of the fuctions in test_fuctions to be able to test of the existence of a file other than journal.txt. I also made it so you can add a particular test file when calling ./test as the second param, requiring the first. This makes the output easier to look at as I add tests for new funcitionality.