Enhancement for https://github.com/greenmanspirit/journal.txt/issues/7
Added new tests and the functionality to make the tests pass. This was pretty simple and you can now set your own EDITOR and FILEDIR in .journalrc and it wil be respected by journal.rb. I also updated the README with information about .journalrc. The emacs script was created to support testing. It does the same as vim but also outputs 'emacs_out' to show that is ran.  I did this so I could set a non-default editor in .journalrc and tell if it ran.