Bug fix for https://github.com/greenmanspirit/journal.txt/issues/23
This was an easy fix, I just added the file that wasn't being cleaned up to my call for cleanup in the 'n' handler for the delete question. I was unable to peproduce it without putting things in the testing scripts to stop it and have it's environment to test in.