For https://github.com/greenmanspirit/journal.txt/issues/26
Removed write entry test from new_test.sh and placed it into journaltxt_test.sh and renamed it entry format. I did this because that was testing a formatting thing and that is more appropriate in journaltxt_test.sh. I also added three more tests to test that the entries are in descending date format.