Bug fix for https://github.com/greenmanspirit/journal.txt/issues/29
So this was interesting. As I posted in the ticket, the tests were not properly testing things. To more properly handle this I expanded the journal.rb script to print out the filter regex if you put a t at the start of the filter, ie t01/*/14. This is because I do not have low level access to the ruby code from bash. I then updated the tests to verify that the regexes are what I expect. Part of this was adding the -F option to grep in test_functions assert_file_contains and assert_file_not_contains so that grep doesn't try to interpret the filter regex. I also added tests for one * in the filter.